+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Added `CT: PlateFromBrep` and `CT: BeamFromBox` GH components.
-* Added `InstallDependencies` GH component for consistent packages versions in the GHPython env.
+* Added `InstallDependencies` GH component for consistent package versions in the GHPython env.
 
 ### Changed